Add special cases for Input being Void
Since `Delegated` is a generic type, it treats `Void` just as any
other type. So when Input is `Void`, you need to write
`.call(())`, which can sometimes be not obvious. This commit
adds special cases for when Input is `Void`, and when both Input
and Output are `Void`, so that the usage code can look nicer

Addresses https://medium.com/@alanjcushway/this-is-pretty-cool-though-the-first-thing-i-tried-was-1579a9974d9